Add endpoint, reorganize v1 endpoints, more...
1. app/main.py

- Update imports

2. app/v1/endpoints/inventory.py

- Populate fabricName when saving to database

3. app/v1/endpoints/lan_fabric/rest_control_switches_fabric_name.py

- Add endpoint

/appcenter/cisco/ndfc/api/v1/lan-fabric/rest/control/switches/{switch_serial_number}/fabric-name

4.  app/v1/endpoints/lan_fabric/rest_control_switches_overview.py

- Rename from app/v1/endpoints/lan_fabric_rest_control_switches_overview.py
